merge: merging the upstream/main changes to my powerful_inspector branch
