+++ lib/src/requests_inspector_widget.dart
@@ -6,7 +6,6 @@ import 'package:requests_inspector/src/request_stopper_editor_dialog.dart';
 import 'package:requests_inspector/src/response_stopper_editor_dialog.dart';
 import 'package:requests_inspector/src/shared_widgets/inspector.dart';
 import '../requests_inspector.dart';
-import 'share_type_enum.dart';
 
 // Helper class for combining two values for a Selector
 // Placed at top-level for better organization
@@ -137,654 +136,3 @@ class RequestsInspector extends StatelessWidget {
     );
   }
 }
-
-class _Inspector extends StatelessWidget {
-  const _Inspector({
-    super.key,
-    GlobalKey<NavigatorState>? navigatorKey,
-  }) : _navigatorKey = navigatorKey;
-
-  final GlobalKey<NavigatorState>? _navigatorKey;
-
-  bool showStopperDialogsAllowed() => _navigatorKey?.currentContext != null;
-
-  @override
-  Widget build(BuildContext context) {
-    return MaterialApp(
-      theme: ThemeData.dark().copyWith(
-        colorScheme: ColorScheme.dark(primary: Colors.grey[800]!),
-      ),
-      home: Scaffold(
-        appBar: _buildAppBar(context),
-        body: _buildBody(),
-        floatingActionButton: _buildShareFloatingButton(),
-      ),
-    );
-  }
-
-  AppBar _buildAppBar(BuildContext context) {
-    final inspectorController = context.read<InspectorController>();
-
-    return AppBar(
-      backgroundColor: Colors.black,
-      title: const Text('Inspector 🕵'),
-      leading: IconButton(
-        onPressed: inspectorController.hideInspector,
-        icon: const Icon(Icons.close),
-        color: Colors.white,
-      ),
-      actions: [
-        Selector<InspectorController, int>(
-          selector: (_, inspectorController) => inspectorController.selectedTab,
-          builder: (context, selectedTab, _) => selectedTab == 0
-              ? TextButton(
-                  onPressed: () => _showAreYouSureDialog(
-                    context,
-                    onYes: inspectorController.clearAllRequests,
-                  ),
-                  child: const Text(
-                    'Clear All',
-                    style: TextStyle(color: Colors.white),
-                  ),
-                )
-              : _RunAgainButton(
-                  key: ValueKey(inspectorController.selectedRequest.hashCode),
-                  onTap: inspectorController.runAgain,
-                ),
-        ),
-        _buildPopUpMenu(inspectorController),
-      ],
-    );
-  }
-
-  Widget _buildPopUpMenu(InspectorController inspectorController) {
-    return PopupMenuButton(
-      icon: const Icon(Icons.more_vert, color: Colors.white),
-      itemBuilder: (context) => [
-        if (showStopperDialogsAllowed())
-          PopupMenuItem(
-            child: InkWell(
-              onTap: () => inspectorController.requestStopperEnabled =
-                  !inspectorController.requestStopperEnabled,
-              child: Row(
-                mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                children: [
-                  const Text('Requests Stopper'),
-                  Selector<InspectorController, bool>(
-                    selector: (_, inspectorController) =>
-                        inspectorController.requestStopperEnabled,
-                    builder: (context, requestStopperEnabled, _) => Switch(
-                      value: requestStopperEnabled,
-                      activeColor: Colors.green,
-                      activeTrackColor: Colors.grey[700],
-                      inactiveThumbColor: Colors.white,
-                      inactiveTrackColor: Colors.grey[700],
-                      onChanged: (value) =>
-                          inspectorController.requestStopperEnabled = value,
-                    ),
-                  ),
-                ],
-              ),
-            ),
-          ),
-        if (showStopperDialogsAllowed())
-          PopupMenuItem(
-            child: InkWell(
-              onTap: () => inspectorController.responseStopperEnabled =
-                  !inspectorController.responseStopperEnabled,
-              child: Row(
-                mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                children: [
-                  const Text('Responses Stopper'),
-                  Selector<InspectorController, bool>(
-                    selector: (_, inspectorController) =>
-                        inspectorController.responseStopperEnabled,
-                    builder: (context, responseStopperEnabled, _) => Switch(
-                      value: responseStopperEnabled,
-                      activeColor: Colors.green,
-                      activeTrackColor: Colors.grey[700],
-                      inactiveThumbColor: Colors.white,
-                      inactiveTrackColor: Colors.grey[700],
-                      onChanged: (value) =>
-                          inspectorController.responseStopperEnabled = value,
-                    ),
-                  ),
-                ],
-              ),
-            ),
-          ),
-      ],
-    );
-  }
-
-  Widget _buildBody() {
-    return Selector<InspectorController, int>(
-      selector: (_, inspectorController) => inspectorController.selectedTab,
-      builder: (context, selectedTab, _) => Column(
-        children: [
-          _buildHeaderTabBar(context, selectedTab: selectedTab),
-          _buildSelectedTab(context, selectedTab: selectedTab),
-        ],
-      ),
-    );
-  }
-
-  Widget _buildHeaderTabBar(BuildContext context, {required int selectedTab}) {
-    final inspectorController = context.read<InspectorController>();
-    return Row(
-      crossAxisAlignment: CrossAxisAlignment.start,
-      children: [
-        _buildTabItem(
-          title: 'All',
-          isSelected: selectedTab == 0,
-          isLeft: true,
-          onTap: () => inspectorController.selectedTab = 0,
-        ),
-        _buildTabItem(
-          title: 'Request Details',
-          isSelected: selectedTab == 1,
-          isLeft: false,
-          onTap: () => inspectorController.selectedTab = 1,
-        ),
-      ],
-    );
-  }
-
-  Widget _buildTabItem({
-    required String title,
-    required bool isSelected,
-    required bool isLeft,
-    required VoidCallback onTap,
-  }) {
-    return Expanded(
-      child: InkWell(
-        onTap: onTap,
-        child: Container(
-          padding: const EdgeInsets.all(12.0),
-          alignment: Alignment.center,
-          decoration: BoxDecoration(
-            color: isSelected ? const Color(0xFF1C1B1F) : Colors.white,
-            borderRadius: isSelected
-                ? null
-                : BorderRadius.only(
-                    bottomRight: isLeft
-                        ? const Radius.circular(12.0)
-                        : const Radius.circular(0.0),
-                    bottomLeft: isLeft
-                        ? const Radius.circular(0.0)
-                        : const Radius.circular(12.0),
-                  ),
-          ),
-          child: Text(
-            title,
-            style: TextStyle(
-              color: isSelected ? Colors.white : Colors.black,
-              fontWeight: isSelected ? FontWeight.w800 : FontWeight.w300,
-            ),
-          ),
-        ),
-      ),
-    );
-  }
-
-  Widget _buildSelectedTab(BuildContext context, {required int selectedTab}) {
-    return selectedTab == 0
-        ? _buildAllRequests(context)
-        : const _RequestDetailsPage();
-  }
-
-  Widget _buildAllRequests(BuildContext context) {
-    final inspectorController = context.read<InspectorController>();
-    return Expanded(
-      child: Selector<InspectorController, List<RequestDetails>>(
-        selector: (_, controller) => controller.requestsList,
-        shouldRebuild: (previous, next) => true,
-        builder: (context, allRequests, _) => allRequests.isEmpty
-            ? const Center(child: Text('No requests added yet'))
-            : ListView.separated(
-                padding:
-                    const EdgeInsets.symmetric(vertical: 6.0, horizontal: 6.0),
-                separatorBuilder: (_, __) => const SizedBox(height: 6.0),
-                itemCount: allRequests.length,
-                itemBuilder: (context, index) {
-                  final request = allRequests[index];
-                  return _RequestItemWidget(
-                    isSelected: inspectorController.selectedRequest == request,
-                    request: request,
-                    onTap: (request) =>
-                        inspectorController.selectedRequest = request,
-                  );
-                },
-              ),
-      ),
-    );
-  }
-
-  Future<void> _showAreYouSureDialog(
-    BuildContext context, {
-    required VoidCallback onYes,
-  }) {
-    return showDialog<void>(
-      context: context,
-      builder: (context) => AlertDialog(
-        title: const Text('Are you sure? 🤔'),
-        content:
-            const Text('This will clear all requests added to the inspector'),
-        actions: [
-          TextButton(
-            child: const Text(
-              'Yes',
-              style: TextStyle(color: Colors.red),
-            ),
-            onPressed: () {
-              Navigator.of(context).pop();
-              onYes();
-            },
-          ),
-          TextButton(
-            onPressed: Navigator.of(context).pop,
-            child: const Text('No', style: TextStyle(color: Colors.green)),
-          ),
-        ],
-      ),
-    );
-  }
-
-  Widget _buildShareFloatingButton() {
-    return Selector<InspectorController, bool>(
-      selector: (_, inspectorController) =>
-          inspectorController.selectedTab == 1 &&
-          inspectorController.selectedRequest != null,
-      builder: (context, showShareButton, _) => showShareButton
-          ? FloatingActionButton(
-              backgroundColor: Colors.black,
-              foregroundColor: Colors.white,
-              child: const Icon(Icons.share),
-              onPressed: () async {
-                final box = context.findRenderObject() as RenderBox?;
-
-                final controller = context.read<InspectorController>();
-                final selectedRequest = controller.selectedRequest!;
-                final isHttp = _isHttp(selectedRequest);
-
-                final shareType =
-                    isHttp ? await _showDialogShareType(context) : null;
-
-                if (shareType == null) return;
-
-                controller.shareSelectedRequest(
-                  sharePositionOrigin: box == null
-                      ? null
-                      : box.localToGlobal(Offset.zero) & box.size,
-                  shareType: shareType,
-                );
-              })
-          : const SizedBox(),
-    );
-  }
-
-  bool _isHttp(RequestDetails selectedRequest) {
-    return selectedRequest.requestMethod == RequestMethod.GET ||
-        selectedRequest.requestMethod == RequestMethod.POST ||
-        selectedRequest.requestMethod == RequestMethod.PUT ||
-        selectedRequest.requestMethod == RequestMethod.PATCH ||
-        selectedRequest.requestMethod == RequestMethod.DELETE;
-  }
-
-  Future<ShareType?> _showDialogShareType(BuildContext context) {
-    return showDialog<ShareType?>(
-      context: context,
-      builder: (context) => AlertDialog(
-        title: const Text('Normal Log or cURL command? 🤔'),
-        content: const Text(
-            'The cURL command is more useful for exporting to Postman or run it again from terminal'),
-        actions: [
-          TextButton(
-            child: const Text(
-              'cURL Command',
-              style: TextStyle(color: Colors.green),
-            ),
-            onPressed: () => Navigator.of(context).pop(ShareType.CurlCommand),
-          ),
-          TextButton(
-            onPressed: () => Navigator.of(context).pop(ShareType.NormalLog),
-            child: const Text(
-              'Normal Log',
-              style: TextStyle(color: Colors.yellow),
-            ),
-          ),
-          TextButton(
-            onPressed: () => Navigator.of(context).pop(ShareType.Both),
-            child: const Text(
-              'Both',
-              style: TextStyle(color: Colors.red),
-            ),
-          ),
-        ],
-      ),
-    );
-  }
-}
-
-class _RunAgainButton extends StatefulWidget {
-  const _RunAgainButton({
-    Key? key,
-    required this.onTap,
-  }) : super(key: key);
-
-  final Future<void> Function() onTap;
-
-  @override
-  _RunAgainButtonState createState() => _RunAgainButtonState();
-}
-
-class _RunAgainButtonState extends State<_RunAgainButton> {
-  bool _isLoading = false;
-
-  @override
-  Widget build(BuildContext context) {
-    return _isLoading
-        ? const Center(
-            child: Padding(
-            padding: EdgeInsets.all(8.0),
-            child: CircularProgressIndicator(color: Colors.white),
-          ))
-        : InkWell(
-            onTap: () {
-              _setBusy();
-              widget.onTap().whenComplete(_setReady);
-            },
-            child: const Row(
-              mainAxisSize: MainAxisSize.min,
-              children: [
-                Text('Run', style: TextStyle(color: Colors.white)),
-                Icon(Icons.play_arrow, color: Colors.white),
-              ],
-            ));
-  }
-
-  void _setBusy() => setState(() => _isLoading = true);
-  void _setReady() => setState(() => _isLoading = false);
-}
-
-class _RequestItemWidget extends StatelessWidget {
-  const _RequestItemWidget({
-    super.key,
-    required bool isSelected,
-    required RequestDetails request,
-    required ValueChanged<RequestDetails> onTap,
-  })  : _isSelected = isSelected,
-        _request = request,
-        _onTap = onTap;
-
-  final bool _isSelected;
-  final RequestDetails _request;
-  final ValueChanged<RequestDetails> _onTap;
-
-  @override
-  Widget build(BuildContext context) {
-    Widget child = ListTile(
-      shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(4.0)),
-      tileColor: _specifyStatusCodeColor(_request.statusCode),
-      leading: Column(
-        mainAxisAlignment: MainAxisAlignment.spaceEvenly,
-        children: [
-          Text(
-            _request.requestMethod.name,
-            style: const TextStyle(
-              fontSize: 18.0,
-              fontWeight: FontWeight.bold,
-            ),
-          ),
-          Text(
-            _request.receivedTime != null
-                ? _calculateDuration(_request.sentTime, _request.receivedTime!)
-                : _extractTimeText(_request.sentTime),
-            style: TextStyle(color: Colors.grey[800]),
-          ),
-        ],
-      ),
-      title: Text(_request.requestName),
-      subtitle: Text(_request.url),
-      trailing: Text(
-        _request.statusCode?.toString() ?? 'Err',
-        style: const TextStyle(
-          fontSize: 16.0,
-          fontWeight: FontWeight.bold,
-        ),
-      ),
-      onTap: () => _onTap(_request),
-    );
-
-    if (_isSelected)
-      child = DecoratedBox(
-        decoration: BoxDecoration(
-          border: Border.all(color: Colors.white, width: 2.0),
-          borderRadius: BorderRadius.circular(4.0),
-        ),
-        child: child,
-      );
-    return Theme(
-      data: Theme.of(context).copyWith(colorScheme: const ColorScheme.light()),
-      child: child,
-    );
-  }
-}
-
-Color? _specifyStatusCodeColor(int? statusCode) {
-  if (statusCode == null) return Colors.red[400];
-  if (statusCode > 399) return Colors.red[400];
-  if (statusCode > 299) return Colors.yellow[400];
-  return Colors.green[400];
-}
-
-class _RequestDetailsPage extends StatelessWidget {
-  const _RequestDetailsPage();
-
-  @override
-  Widget build(BuildContext context) {
-    return Expanded(
-      child: Selector<InspectorController, RequestDetails?>(
-        selector: (_, inspectorController) =>
-            inspectorController.selectedRequest,
-        shouldRebuild: (previous, next) => true,
-        builder: (context, selectedRequest, _) => selectedRequest == null
-            ? const Center(
-                child: Text('Please select a request first to view details'))
-            : _buildRequestDetails(context, selectedRequest),
-      ),
-    );
-  }
-
-  Widget _buildRequestDetails(BuildContext context, RequestDetails request) {
-    return ListView(
-      padding: const EdgeInsets.only(bottom: 96.0),
-      children: [
-        _buildRequestNameAndStatus(
-          method: request.requestMethod,
-          requestName: request.requestName,
-          statusCode: request.statusCode,
-        ),
-        _buildRequestSentTimeAndDuration(
-          request.sentTime,
-          request.receivedTime,
-        ),
-        _buildTitle('URL:'),
-        _buildSelectableText(request.url),
-        ..._buildHeadersBlock(request.headers),
-        ..._buildQueryBlock(request.queryParameters),
-        ..._buildRequestBodyBlock(request.requestBody),
-        ..._buildResponseBodyBlock(request.responseBody),
-      ].mapIndexed(_buildBackgroundColor).toList(),
-    );
-  }
-
-  Widget _buildBackgroundColor(index, item) {
-    Widget child = Padding(
-      padding: const EdgeInsets.symmetric(vertical: 8.0),
-      child: item,
-    );
-    return index.isEven
-        ? child
-        : Padding(
-            padding: const EdgeInsets.symmetric(horizontal: 6.0),
-            child: DecoratedBox(
-              decoration: const BoxDecoration(
-                color: Color.fromARGB(255, 19, 19, 19),
-                borderRadius: BorderRadius.all(Radius.circular(4.0)),
-              ),
-              child: child,
-            ),
-          );
-  }
-
-  Widget _buildRequestNameAndStatus({
-    RequestMethod? method,
-    String? requestName,
-    int? statusCode,
-  }) {
-    return Padding(
-      padding: const EdgeInsets.all(6.0),
-      child: Row(
-        crossAxisAlignment: CrossAxisAlignment.end,
-        children: [
-          Expanded(
-            child: Text(
-              _createRequestName(method, requestName),
-              style: const TextStyle(
-                fontSize: 16.0,
-                fontWeight: FontWeight.bold,
-              ),
-            ),
-          ),
-          Container(
-            padding: const EdgeInsets.all(6.0),
-            decoration: BoxDecoration(
-              borderRadius: BorderRadius.circular(8.0),
-              color: _specifyStatusCodeColor(statusCode),
-            ),
-            child: Text(
-              statusCode?.toString() ?? 'Err',
-              style: const TextStyle(fontSize: 16.0),
-            ),
-          ),
-        ],
-      ),
-    );
-  }
-
-  String _createRequestName(RequestMethod? method, String? requestName) {
-    return (method?.name == null ? '' : '${method!.name}: ') +
-        (requestName ?? 'No name');
-  }
-
-  Widget _buildRequestSentTimeAndDuration(
-    DateTime sentTime,
-    DateTime? receivedTime,
-  ) {
-    final sentTimeText = _extractTimeText(sentTime);
-    var text = 'Sent at: $sentTimeText';
-
-    if (receivedTime != null) {
-      final durationText = _calculateDuration(sentTime, receivedTime);
-      final receivedTimeText = _extractTimeText(receivedTime);
-      text += '\nReceived at: $receivedTimeText\nDuration: $durationText';
-    }
-
-    return Padding(
-      padding: const EdgeInsets.all(6.0),
-      child: Text(
-        text,
-        style: const TextStyle(fontSize: 16.0),
-      ),
-    );
-  }
-
-  Iterable<Widget> _buildHeadersBlock(headers) {
-    if (headers == null) return [];
-    if ((headers is Map || headers is String || headers is List) &&
-        headers.isEmpty) return [];
-
-    return [
-      _buildTitle('Headers:'),
-      _buildSelectableText(headers),
-    ];
-  }
-
-  Iterable<Widget> _buildQueryBlock(queryParameters) {
-    if (queryParameters == null) return [];
-    if ((queryParameters is Map ||
-            queryParameters is String ||
-            queryParameters is List) &&
-        queryParameters.isEmpty) return [];
-
-    return [
-      _buildTitle('Parameters:'),
-      _buildSelectableText(queryParameters),
-    ];
-  }
-
-  Iterable<Widget> _buildRequestBodyBlock(requestBody) {
-    if (requestBody == null) return [];
-    if ((requestBody is Map || requestBody is String || requestBody is List) &&
-        requestBody.isEmpty) return [];
-
-    return [
-      _buildTitle('RequestBody:'),
-      _buildSelectableText(requestBody),
-    ];
-  }
-
-  Iterable<Widget> _buildResponseBodyBlock(responseBody) {
-    if (responseBody == null) return [];
-    if ((responseBody is Map ||
-            responseBody is String ||
-            responseBody is List) &&
-        responseBody.isEmpty) return [];
-    return [
-      _buildTitle('ResponseBody:'),
-      _buildSelectableText(responseBody),
-    ];
-  }
-
-  Widget _buildSelectableText(text) {
-    final prettyprint = JsonPrettyConverter().convert(text);
-
-    return Padding(
-      padding: const EdgeInsets.all(6.0),
-      child: SelectableText(prettyprint),
-    );
-  }
-
-  Widget _buildTitle(String title) {
-    return Padding(
-      padding: const EdgeInsets.fromLTRB(6.0, 8.0, 6.0, 0.0),
-      child: Text(
-        title,
-        style: const TextStyle(
-          fontSize: 16.0,
-          fontWeight: FontWeight.bold,
-        ),
-      ),
-    );
-  }
-}
-
-String _extractTimeText(DateTime sentTime) {
-  var sentTimeText = sentTime.toIso8601String().split('T').last.substring(0, 8);
-  sentTimeText = _replaceLastSeparatorWithDot(sentTimeText);
-  return sentTimeText;
-}
-
-String _calculateDuration(DateTime sentTime, DateTime receivedTime) {
-  final duration = receivedTime.difference(sentTime);
-
-  if (duration.inMilliseconds < 1000) return '${duration.inMilliseconds} ms';
-  if (duration.inSeconds < 60) return '${duration.inSeconds} s';
-  if (duration.inMinutes < 60) return '${duration.inMinutes} m';
-  if (duration.inHours < 24) return '${duration.inHours} h';
-  return '${duration.inDays} d';
-}
-
-String _replaceLastSeparatorWithDot(String sentTimeText) =>
-    sentTimeText.replaceFirst(':', '.', 5);
